| test env                  | before | after  |
| ------------------------- | ------ | ------ |
| test (ubuntu-latest, 3.7) | 3m 55s | 1m 17s |
| test (ubuntu-latest, 3.8) | 1m 37s | 1m 22s |
| test (ubuntu-latest, 3.9) | 6m 28s | 1m 32s |
| test (ubuntu-latest, 3.10) | 1m 36s | 1m 48s |
| test (ubuntu-latest, pypy-3.8) | 4m 17s | 1m 49s |
| test (macos-latest, 3.7) | 2m 5s | 1m 37s |
| test (macos-latest, 3.8) | 2m 17s | 1m 45s |
| test (macos-latest, 3.9) | 2m 55s | 1m 51s |
| test (macos-latest, 3.10) | 2m 28s | 1m 59s |
| test (windows-latest, 3.7) | 3m 7s | 3m 25s |
| test (windows-latest, 3.8) | 2m 31s | 2m 37s |
| test (windows-latest, 3.9) | 3m 58s | 3m 36s |
| test (windows-latest, 3.10) | 3m 57s | 2m 24s |
| test (windows-latest, pypy-3.7) | 5m 53s | 3m 39s |

big differences usually due to super-slow pip install, but unit test time improved 25-50%